minor, make con6ity separable (and separate it)
First and foremost, this changeset is not strictly necessary, it's more
of a clean-up.  con6ity.c gets a companion header to be even more decoupled
from libev and network mumbo jumbo.

Makefile.am is changed to demonstrate the separate build but this will be
reverted in the next commit.

handle_close() and handle_data() have become weak functions with some nop
defaults in con6ity.

Signed-off-by: Sebastian Freundt <sebastian.freundt@ga-group.nl>