fix, interfere with libxml parser when in glue mode
This changeset provides the magic to `stop' and `resume' libxml's push parser.
The magic begins when we enter the glue mode (pushing UMPF_TAG_GLUE onto the
context state stack).  Since the parser is then still active our characters
callback will be called upon the next run, in there we call our own push
routine which goes on the prowl for a matching end tag itself, when found
libxml's internal buffers are fiddled with to make it look as though the push
parser has done an excellent job.
When the end tag is not found in the first go every call to parse_more_blob*()
will call our version of the stuff buf pusher, libxml won't be served any
such buffer.  When finally the end tag is found, glue mode is left and
parsing continues as per usual.

The uglicity of this approach makes me wonder if we need our own sax-like
xml parser.  libxml2 does a poor job (memory- and performance-wise) in all
scenarios so far.

Signed-off-by: Sebastian Freundt <sebastian.freundt@ga-group.nl>