minor, provide a dedicated structure for instruments
This changeset prepares SecDef* support where the Instrmt tag is used as well.
To unify its usage, put everything that makes up an instrument into a dedicated
struct, actually only the info used to identify the instrument is in there,
due to our close intertanglement with FIXML and for efficiency reasons.

Whilst at it, the struct ins_qty is cleaned up, Xqty_s is renamed to qty_s.

Signed-off-by: Sebastian Freundt <hroptatyr@fresse.org>