fix, HACK, find glue end tag even in incomplete buffers
This changeset peruses the in buffer beyond its designated length for the end-of-glue tag.
This is useful when the end-of-glue tag is split across buffers but may not work in general
for instance when we specifically only pass "</aou:" to the push parser, there's no way
ever for this hack to work.

Signed-off-by: Sebastian Freundt <hroptatyr@fresse.org>